move dotted-divider into care-ui location
